requirements.txt
altair==5.2.0
annotated-types==0.6.0
ansi2html==1.9.1
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
autopep8==2.0.4
Babel==2.14.0
backoff==2.2.1
beautifulsoup4==4.12.2
black==24.4.0
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dateutils==0.6.12
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.8
duckdb==0.10.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==2.0.1
Faker==28.0.0
fastjsonschema==2.19.1
favicon==0.7.0
filelock==3.15.1
Flask==3.0.0
fonttools==4.47.0
fqdn==1.5.1
gitdb==4.0.11
GitPython==3.1.40
gotrue==2.4.2
greenlet==3.0.3
gw_dsl_parser==0.1.47
h11==0.14.0
htbuilder==0.6.2
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.6
importlib-metadata==6.11.0
iniconfig==2.0.0
ipykernel==6.28.0
ipython==8.19.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itables==1.6.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.1
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.6.0
jupyter_contrib_core==0.4.2
jupyter_contrib_nbextensions==0.7.0
jupyter_core==5.7.0
jupyter_server==2.12.2
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kanaries_track==0.0.4
kiwisolver==1.4.5
lxml==5.0.1
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
monotonic==1.6
more-itertools==10.4.0
MouseInfo==0.1.3
mplsoccer==1.2.4
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==6.4.12
notebook_shim==0.2.3
numpy==1.26.3
openpyxl==3.1.5
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
pillow==10.2.0
pipenv==2024.0.1
platformdirs==4.1.0
plotly==5.18.0
pluggy==1.5.0
postgrest==0.16.2
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.7
psycopg2==2.9.9
pure-eval==0.2.2
pyarrow==14.0.2
PyAutoGUI==0.9.54
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
PyGetWindow==0.0.9
Pygments==2.17.2
pygwalker==0.4.8
pymdown-extensions==10.9
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytest==8.3.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytweening==1.2.0
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
realtime==1.0.3
referencing==0.32.1
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.16.2
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
segment-analytics-python==2.2.3
selenium==4.19.0
Send2Trash==1.8.2
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
sqlglot==23.10.0
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
storage3==0.7.4
streamlit==1.37.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-elements==0.1.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.7
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
StrEnum==0.4.15
supabase==2.4.1
supafunc==0.4.5
tenacity==8.2.3
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.4
traitlets==5.9.0
trio==0.25.0
trio-websocket==0.11.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uri-template==1.3.0
urllib3==2.1.0
validators==0.22.0
virtualenv==20.26.2
wasmtime==12.0.0
watchdog==3.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wsproto==1.2.0
XlsxWriter==3.1.9
zipp==3.17.0
